Basic addr2line cachine is working... needs more verification
We were running into an issue where any use of the junction map data
structures would cause our exec function to never return. This was
actually triggered by us not unsetting the LD_PRELOAD env var before
calling popen. It is a bit of a mystery why using the junction functions
triggered this issue, as they seem somewhat independent. Nonetheless, in
the commit 'make test' is working as planned. 'make log' seems to seg
fault in some sitations, we should investigate this.