Fixed bug with some condition variables not logging correctly.
This was because the init function is only called once (I believe). So
previously, the use_real_func flag defaulted to true, and was never set
to false (normally done so in init). Instead, I set it to false to begin
with. However, this means that there could be interpositioning that
occurs before our initializer is called...