TSX updated and works, however, aborts too often
The test on pthread_test seem promising and predictable. Low number of
threads leads to low contention, and the transaction succeeds most of
the time. However, this is simply not the case with rocksDB. Even with
only one thread, the transaction falls back to the spin lock about every
time. And none of the assosciated error flags are set...